chore: a bit of code cleanup
